+++ src/States/InteractiveState.h
@@ -21,7 +21,7 @@ namespace Raycer
 		virtual void update(double timeStep) = 0;
 		virtual void render(double timeStep, double interpolation) = 0;
 
-		virtual void windowResized(size_t width, size_t height) = 0;
-		virtual void framebufferResized(size_t width, size_t height) = 0;
+		virtual void windowResized(int width, int height) = 0;
+		virtual void framebufferResized(int width, int height) = 0;
 	};
 }
